converter_app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
filetype==1.2.0
Pillow==10.4.0
openpyxl==3.1.5
jinja2==3.1.4